Use headers instead of body to pass the parameters

+++ src/endpoints/GenerateReport.ts
@@ -1,5 +1,6 @@
 import http from 'http';
 import { MainDB } from '../DataBase/MainDB.ts';
+import { Utils } from '../Utils.ts';
 
 
 export class GenerateReport
@@ -11,13 +12,25 @@ export class GenerateReport
     {
         res.statusCode = 200;
 
-        MainDB.PurgeUsers();
-        res.write(JSON.stringify({
-            active_users: MainDB.ActiveUsers.Size(),
-            active_versions: MainDB.ActiveVersions.GetReport_ByShareMap(),
-            active_managers: MainDB.ActiveManagers.GetReport_ByBitMask(),
-            program_ranking: MainDB.InstallsRanking.GetProgramRanking(10),
-        }))
+        const apiKey = Utils.GetHeader(req, 'apiKey');
+        if(apiKey == "")
+        {
+            res.statusCode = 406;
+        }
+        else if (!Utils.Authenticate(apiKey))
+        {
+            res.statusCode = 403;
+        }
+        else
+        {
+            MainDB.PurgeUsers();
+            res.write(JSON.stringify({
+                active_users: MainDB.ActiveUsers.Size(),
+                active_versions: MainDB.ActiveVersions.GetReport_ByShareMap(),
+                active_managers: MainDB.ActiveManagers.GetReport_ByBitMask(),
+                program_ranking: MainDB.InstallsRanking.GetProgramRanking(10),
+            }))
+        }
     }
 }
